api/requirements.txt
mangum
requests
fastapi
python-jose
mypy-boto3
mypy-boto3-dynamodb